More cleanup, added way to recreate thumbnails Lots and lots of cleanup, including fixing a nasty bug whereh I forgot to init my RECTs after converting them from LPRECTs, and type safety imrprovements. Added a menu option to the settings window that allows users to recreate thumbnails if they change DPI or something.

+++ Winclipper/Bitmap.h
@@ -16,7 +16,7 @@ public:
 	const std::vector<RGBQUAD>					RgbQuadCollection();
 	const std::shared_ptr<BYTE>					DibBitmapBits() noexcept;
 	HBITMAP										GetHbitmap();
-	size_t										Size() noexcept;
+	DWORD										Size() noexcept;
 
 private:
 	std::shared_ptr<BITMAPINFOHEADER>			pDibBitmapInfoHeader = nullptr;
